+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# [Unreleased-main] - 2026-05-01
+# [Unreleased-main] - 2026-05-13
 
 ## Added
 - Electricity consumption calculation of geothermal assets, using the defined COP. 
@@ -9,7 +9,8 @@
 - First implementation of a temperature profile for a producer. (experimental)
 - Test is added to check airco sizing and cost calculation
 - Workflow of EndScenarioSizing with demand uncertainty is added.
-- Maximum and minimum temperature of residual heat source are parsed from esdl
+- Ramp constraints for heat producers are added.
+- Maximum and minimum temperature of heat sources are parsed from esdl
 
 ## Changed
 - Reduced the number of constraints required for headloss calculation with LINEARIZED_N_LINES_EQUALITY setting.
@@ -19,6 +20,8 @@
 - Updated pyESDL to v26.3
 - Updated the GROW workflow to provide feedback on which heat demands are not matched in priority 1.
 - Removed old test files that are no longer used.
+- Added electricity cost profile to variable operational cost of assets which convert electricity to heat.
+- Heating_and_cooling example is cleaned up
 
 ## Fixed
 - ProfileConstraints: Use already available function to get profile quantity and unit